separate Hamburger and Nav from Header
